tests: split out match tests and validate alternative algorithms
